[MIL-1] refactor: move config files
