+++ config/random_statements.js
