+++ lib/app.js
@@ -4,11 +4,11 @@ const bodyParser = require('body-parser');
 const enforce = require('express-sslify');
 const app = express();
 
-const error_messages = require('./config/error_messages');
+const { ENV_MISSING } = require('../config/error_messages');
 
 const enableHttps = () => {
     if (!process.env.ENV) {
-        throw error_messages.ENV_MISSING;
+        throw ENV_MISSING;
     }
     const enableScript = process.env.ENV === 'local' ? null : app.use(enforce.HTTPS({ trustProtoHeader: true }));
     return enableScript;
