+++ lib/controller.js
@@ -3,9 +3,9 @@ const _ = require('lodash');
 const axios = require('axios');
 
 // config related variables
-const error_messages = require('./config/error_messages');
-const { statements, mysteryStatements } = require('./config/statements');
-const { GITHUB_API_URL } = require('./config/api');
+const { CLIENT_ID_AND_CLIENT_SECRET_MISSING } = require('../config/error_messages');
+const { statements, mysteryStatements } = require('../config/random_statements');
+const { GITHUB_API_URL } = require('../config/api_url');
 
 // load utils
 const { emojiGenerator } = require('./public/js/emoji_generator');
@@ -15,7 +15,7 @@ const { getColorsForLanguages } = require('./public/js/language_colors_generator
 const client_id = process.env.CLIENT_ID || false;
 const client_secret = process.env.CLIENT_SECRET || false;
 if (!(client_id && client_secret)) {
-    throw error_messages.CLIENT_ID_AND_CLIENT_SECRET_MISSING;
+    throw CLIENT_ID_AND_CLIENT_SECRET_MISSING;
 }
 const clientParams = `client_id=${client_id}&client_secret=${client_secret}`;
